refactor: move agent-session to lib/infra/
Part of #7697 (restructure src/lib/ into layered architecture).
Move lib/agent-session/ to lib/infra/agent-session/ since this module
manages session CRUD and is pure infrastructure.

Closes #7722
