+++ turbo/apps/web/app/api/webhooks/agent/complete/__tests__/route.test.ts
@@ -19,7 +19,7 @@ import {
 } from "../../../../../../src/__tests__/api-test-helpers";
 import { enqueueRun } from "../../../../../../src/lib/zero/zero-run-queue-service";
 import { reloadEnv } from "../../../../../../src/env";
-import { getAgentSessionWithConversation } from "../../../../../../src/lib/agent-session";
+import { getAgentSessionWithConversation } from "../../../../../../src/lib/infra/agent-session";
 import {
   testContext,
   uniqueId,
