+++ turbo/apps/web/src/lib/checkpoint/checkpoint-service.ts
@@ -4,7 +4,7 @@ import { agentComposeVersions } from "../../db/schema/agent-compose";
 import { conversations } from "../../db/schema/conversation";
 import { checkpoints } from "../../db/schema/checkpoint";
 import { notFound } from "../errors";
-import { createAgentSession, updateAgentSession } from "../agent-session";
+import { createAgentSession, updateAgentSession } from "../infra/agent-session";
 import { storeSessionHistory } from "../session-history";
 import { logger } from "../logger";
 import type {
